+++ CHANGELOG.md
@@ -8,9 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 Unreleased in the current development version (target v0.19.0):
 
 AQUA core complete list:
+- Timeseries graphical function adapted to plot multiple levels for ocean diagnostic (#2328)
 - Extending evaluate formula method to exponential and parenthesis (#2327)
 
 AQUA diagnostics complete list:
+- Ocean Drift: timeseries plotting function with update in cli (#2322)
 - Gregory: more detailed description and plot labels (#2306)
 
 ## [v0.18.0]